+++ chat-widget/src/components/webchatcontainerstateful/webchatcontroller/middlewares/renderingmiddlewares/attachments/MaliciousAttachment.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import DownloadBlockedAttachment from "./DownloadBlockedAttachment";
+import { CrossIcon, MaliciousFileIcon } from "../../../../../../assets/Icons";
+
+const MaliciousAttachment = (props: any) => { // eslint-disable-line @typescript-eslint/no-explicit-any
+    const extraContent = <img src={CrossIcon} alt="Malicious" />;
+    return (
+        <DownloadBlockedAttachment {...props} iconData={MaliciousFileIcon} extraContent={extraContent}/>
+    );
+};
+
+export default MaliciousAttachment;
